Stock_Analyzer/select_sector/data_scientist.py
# select_sector/data_scientist.py

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split

class SectorDataScientist:
    def __init__(self, historical_data):
        self.historical_data = historical_data
        self.model = RandomForestClassifier()

    def prepare_data(self):
        # Placeholder for preparing the dataset
        data = []
        for company, df in self.historical_data.items():
            df['Returns'] = df['Close'].pct_change()
            df.dropna(inplace=True)
            data.append(df[['Close', 'Returns']].reset_index())
        return pd.concat(data)

    def train_model(self, X, y):
        self.model.fit(X, y)

    def predict_top_stocks(self, X):
        probabilities = self.model.predict_proba(X)[:, 1]
        return probabilities

if __name__ == "__main__":
    # Placeholder for historical data
    historical_data = {'AAPL': pd.DataFrame(), 'GOOGL': pd.DataFrame()}  # Example data
    scientist = SectorDataScientist(historical_data)
    prepared_data = scientist.prepare_data()
    
    # Example: Split the data and train the model
    X = prepared_data.drop('Returns', axis=1)
    y = (prepared_data['Returns'] > 0).astype(int)  # Binary classification
    scientist.train_model(X, y)
    
    # Get predictions
    probabilities = scientist.predict_top_stocks(X)
    print(probabilities)
